continue docs just to see order

Index

BOOP.expected_improvement

BOOP.first_func

BOOP.upper_confidence_bound

Ramaining functions

BOOP.expected_improvement — Method

  expected_improvement(gp, xnew, ybest; ξ = 0.01)

Computes the expected improvement given a Gaussian process (gp) object, and the earliest best evaluation (ybest) at the new evaluation point xnew). ξ is a tuning parameter that controls the exploration-exploitation trade-off. A large value of ξ encourages exploration and vice versa.

Returns the expected improvement at xnew.

Examples

julia> Set up GP model.
julia> X_train = [1.0, 2.5, 4.0]; julia> y_train = [sin(x) for x in X_train];
julia> gp_model = GP(X_train', y_train, MeanZero(), SE(0.0, 0.0));
julia> optimize!(gp_model);

julia> # 2. Define the best observed value and a candidate point
julia> y_best = minimum(y_train);
julia> x_candidate = 3.0;

julia> # 3. Compute Expected Improvement
julia> ei = expected_improvement(gp_model, x_candidate, y_best; ξ=0.01)
≈ 4.89.

BOOP.first_func — Method

    first_func(x)

A first dummy function that returns the log of the square of the input value.

julia> T = 500; ϕ = [sin.(2π*(1:T)/T) -0.1*ones(T)]; θ = zeros(T); σ = 1; μ = 2;
julia> y = simTvARMA(ϕ, θ, μ, σ);
julia>  tvPeriodogram(y, 25, 15)

BOOP.upper_confidence_bound — Method

  upper_confidence_bound(gp, xnew; κ = 2.0)

Computes the upper confidence bound criteria at the new point xnew given a Gaussian process (gp) object, and the exploitation/exploitation parameter κ. High values of κ encourage exploration.

Returns the upper confidence bound criteria at xnew.

Examples

julia> Set up GP model.
julia> X_train = [1.0, 2.5, 4.0]; julia> y_train = [sin(x) for x in X_train];
julia> gp_model = GP(X_train', y_train, MeanZero(), SE(0.0, 0.0));
julia> optimize!(gp_model);
julia> x_candidate = 3.0;

julia> ucb = upper_confidence_bound(gp_model, x_candidate, κ = 2.0)
-0.22727575567547253